requirements.txt
# Core Dependencies
pydantic
python-dotenv
requests
google-genai
groq
pinecone
sarvamai
twilio
# Data & Visualisation
streamlit
pandas
plotly
paho-mqtt
fastapi
uvicorn
numpy
python-multipart
# Database & OCR
psycopg2-binary
asyncpg
pytesseract
pillow
boto3
pypdf
# Local Development
# ollama (handled via direct API)
